backend_backup_phase2_2026-01-02_0906/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
psycopg[binary]==3.2.3
SQLAlchemy>=2.0
alembic>=1.13
psycopg[binary]>=3.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
pydantic-settings>=2.2
email-validator>=2.1
argon2-cffi>=23.1.0
passlib[argon2]>=1.7.4
